pyproject.toml
[tool.poetry]
name = "dtms-drexel-bot"
version = "0.1.0"
description = ""
authors = ["Shahriyar Shawon <ShahriyarShawon321@gmail.com>"]
readme = "README.md"
packages = [{include = "dtms_drexel_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.2.2"
dtms-client = "^0.1.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.3.0"
pylint = "^2.17.1"
mypy = "^1.1.1"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
